+++ src/main/java/plugily/projects/villagedefense/handlers/reward/Reward.java
@@ -98,8 +98,7 @@ public class Reward {
 
   public enum RewardType {
     END_GAME("endgame"), END_WAVE("endwave"), ZOMBIE_KILL("zombiekill"),
-    PLAYER_DEATH("playerDeath"), VILLAGER_DEATH("villagerDeath"),
-    SCOREBOARD_REMOVED("scoreboard_remove"), SCOREBOARD_ADD("scoreboard_add");
+    PLAYER_DEATH("playerDeath"), VILLAGER_DEATH("villagerDeath");
 
     private final String path;